refactor(scene): extract the celestial disc + night-star helpers into background/celestial.rs (#469)
Move-only, behavior-preserving. background/mod.rs had grown to ~1962 lines
carrying three paint concerns (window/skyline structure, weather particles, and
the celestial disc + star field). The celestial code is a self-contained unit
(own consts, own position/twinkle hash noise, no shared mutable state), so it
gets its own file like sky.rs / lighting.rs / glass.rs.

Moved verbatim (only visibility -> pub(super) + explicit imports): struct Disc,
compute_disc, night_star_strength, star_exists, star_twinkle, golden_hour_blaze,
and the DISC_/GLOW_/MOON_SHADOW/FIRST_WINDOW_X/HORIZON_/ARC_/MIN_DISC_VIS/STAR_
consts. mod.rs adds `mod celestial;` + an internal `use celestial::*;` (NOT
re-exported up to pixel_painter -- no other pass reads them; the sky branch of
paint_floor_to_ceiling_window + its render tests are the only consumers). The
render-integration disc/star tests stay in mod.rs::tests with the painter they
exercise, reaching the moved symbols through the same re-import.

mod.rs 1962 -> 1776; celestial.rs is 204 lines.

Proof of no behavior change:
- pixtuoid-scene lib tests 496/496 (exact baseline), clippy clean, `just arch` green
- 5 deterministic snapshots (dawn/dusk disc, night stars x2, overcast-hidden)
  byte-IDENTICAL before vs after (shasum)

Docs-currency: theme/mod.rs + sky.rs + scene CLAUDE.md location refs updated to
background/celestial.rs.
